requirements.txt
numpy>=1.18.4
opencv-python>=4.2.0.34
Pillow>=7.0.0
pytesseract>=0.3.4
